derive depot from safes instead of making it separate resource (#2074)
